backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "siteforge-backend"
version = "0.1.0"
description = "SiteForge backend — FastAPI + PostgreSQL for telecom TSSR/BOQ documentation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.18",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]
